+++ CHANGELOG.md
@@ -869,9 +869,16 @@
 - Use a single comment model for all comments
 - Track the creation of adjustments in an activity's Change history
 
+## [release-81] 2021-10-14
+
+- Record changes to refunds and show them in the activity's Change history
+- Show the commitment value on an activity's finance tab
+- Include any commitment in an activity's IATI XML
+
 ## [unreleased]
 
-[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-80...HEAD
+[unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-81...HEAD
+[release-81]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-80...release-81
 [release-80]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-79...release-80
 [release-79]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-78...release-79
 [release-78]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-77...release-78